#12 同じ操作を担当するファイルがあったため不要なファイルを削除などした - 正しい場所へディレクトリ構造を変更したため、importを修正も同時に実装
