+++ pages/RegisterMember/InformCompletedRegistrationMember.tsx
