+++ pages/RegisterMember/InformCompletedTempRegistrationMember.tsx
